tls: deprecate undocumented internal modules
