+++ lib/tls.js
@@ -46,8 +46,8 @@ const EventEmitter = require('events');
 const { URL } = require('internal/url');
 const DuplexPair = require('internal/streams/duplexpair');
 const { canonicalizeIP } = internalBinding('cares_wrap');
-const _tls_common = require('_tls_common');
-const _tls_wrap = require('_tls_wrap');
+const _tls_common = require('internal/tls/common');
+const _tls_wrap = require('internal/tls/wrap');
 
 // Allow {CLIENT_RENEG_LIMIT} client-initiated session renegotiations
 // every {CLIENT_RENEG_WINDOW} seconds. An error event is emitted if more
